+++ package.json
@@ -1,23 +1,10 @@
 {
   "name": "personalwebsite-gatsby",
   "description": "my personal website",
-  "version": "0.1.0",
+  "version": "1.1.0",
   "author": "Luís Sousa <luis.srl.sousa@gmail.com>",
   "dependencies": {
-    "babel-plugin-styled-components": "^1.13.3",
-    "gatsby": "^3.14.6",
-    "gatsby-cli": "^3.14.2",
-    "gatsby-image": "^3.11.0",
-    "gatsby-plugin-google-analytics": "^4.15.0",
-    "gatsby-plugin-manifest": "^3.14.0",
-    "gatsby-plugin-offline": "^4.14.0",
-    "gatsby-plugin-react-helmet": "^4.14.0",
-    "gatsby-plugin-sharp": "^3.14.3",
-    "gatsby-plugin-styled-components": "^4.14.0",
-    "gatsby-source-filesystem": "^3.14.0",
-    "gatsby-transformer-json": "^3.14.0",
-    "gatsby-transformer-remark": "^3.2.0",
-    "gatsby-transformer-sharp": "^3.14.0",
+    "gatsby": "^4.0.0",
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -25,9 +12,7 @@
     "react-icons": "^4.3.1",
     "react-scroll": "^1.8.7",
     "react-tsparticles": "^1.43.1",
-    "react-typed": "^1.2.0",
-    "styled-components": "^5.3.5",
-    "wowjs": "^1.1.3"
+    "typewriter-effect": "^2.18.2",
   },
   "devDependencies": {
     "prettier": "2.2.1"
