+++ src/pages/index.js
@@ -1,18 +1,20 @@
 import React from "react";
-import Layout from "../components/layout";
-import Home from "../components/home";
-import About from "../components/about";
-import Experience from "../components/experience";
-import Education from "../components/education";
-import Projects from "../components/projects";
-import Skills from "../components/skills";
-import Contact from "../components/contact";
-import Footer from "../components/footer";
+import {
+  Layout,
+  Home,
+  About,
+  Experience,
+  Education,
+  Skills,
+  Projects,
+  Contact,
+  Footer
+} from "./components/index";
 
 const IndexPage = () => {
   return (
     <Layout>
-      <Home />
+      {/* <Home /> */}
       <About />
       <Experience />
       <Education />
